[Decomp][Chat] Cohesion split: extract argument extraction to ChatArgExtract.cpp
